Reorganize components for Mapping view
