+++ client/src/views/Mapping/components/TabColumnSuggestion/index.tsx
@@ -1,0 +1,5 @@
+import * as React from "react";
+
+const TabColumnSuggestion = () => <div>Suggested columns</div>;
+
+export default TabColumnSuggestion;
