--- client/src/views/Mapping/components/tabs/columnSuggestionTab.tsx
@@ -1,5 +1,0 @@
-import * as React from "react";
-
-const ColumnSuggestionTab = () => <div>Suggested columns</div>;
-
-export default ColumnSuggestionTab;
